requirements.txt
PyMetaTrader5
playsound
pystray
Pillow
pywin32